added-score_up-in-stack,mal,pokemon
